requirements.txt
streamlit
tensorflow>=2.16.0
numpy
pillow
gdown